Apply suggestions from code review

+++ deepmd/dpmodel/atomic_model/dp_atomic_model.py
@@ -240,7 +240,7 @@ class DPAtomicModel(BaseAtomicModel):
         """Check if the model has default frame parameters."""
         return self.fitting.has_default_fparam()
 
-    def get_default_fparam(self) -> list[int] | None:
+    def get_default_fparam(self) -> list[float] | None:
         """Get the default frame parameters."""
         return self.fitting.get_default_fparam()
